Split canvas.py into subpackage
