+++ core/src/toga/widgets/canvas/__init__.py
@@ -1,0 +1,54 @@
+from .canvas import Canvas, OnResizeHandler, OnTouchHandler
+from .context import ClosedPathContext, Context, FillContext, StrokeContext
+from .drawingobject import (
+    Arc,
+    BeginPath,
+    BezierCurveTo,
+    ClosePath,
+    DrawingObject,
+    Ellipse,
+    Fill,
+    LineTo,
+    MoveTo,
+    QuadraticCurveTo,
+    Rect,
+    ResetTransform,
+    Rotate,
+    Scale,
+    Stroke,
+    Translate,
+    WriteText,
+)
+from .geometry import arc_to_bezier, sweepangle
+
+__all__ = [
+    "Canvas",
+    "OnResizeHandler",
+    "OnTouchHandler",
+    # Drawing Objects
+    "Arc",
+    "BeginPath",
+    "BezierCurveTo",
+    "ClosePath",
+    "DrawingObject",
+    "Ellipse",
+    "Fill",
+    "LineTo",
+    "MoveTo",
+    "QuadraticCurveTo",
+    "Rect",
+    "ResetTransform",
+    "Rotate",
+    "Scale",
+    "Stroke",
+    "Translate",
+    "WriteText",
+    # Context
+    "ClosedPathContext",
+    "Context",
+    "FillContext",
+    "StrokeContext",
+    # Geometry,
+    "arc_to_bezier",
+    "sweepangle",
+]
